requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
Django==5.2.1
django-admin-list-filter-dropdown==1.0.3
django-filter==25.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
greenlet==3.2.2
playwright==1.52.0
pyee==13.0.0
soupsieve==2.7
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.13.2
